mov s_mp_rand_source_platform to separate file
