+++ callgraph.txt
@@ -15360,6 +15360,9 @@ BN_S_MP_MUL_HIGH_DIGS_FAST_C
 +--->BN_MP_CLAMP_C
 
 
+BN_S_MP_RAND_SOURCE_PLATFORM_C
+
+
 BN_S_MP_REVERSE_C
