+++ tommath_class.h
@@ -150,6 +150,7 @@
 #   define BN_S_MP_MUL_DIGS_FAST_C
 #   define BN_S_MP_MUL_HIGH_DIGS_C
 #   define BN_S_MP_MUL_HIGH_DIGS_FAST_C
+#   define BN_S_MP_RAND_SOURCE_PLATFORM_C
 #   define BN_S_MP_REVERSE_C
 #   define BN_S_MP_SQR_C
 #   define BN_S_MP_SQR_FAST_C
@@ -720,6 +721,8 @@
 #   define BN_MP_DIV_2_C
 #   define BN_MP_MUL_2_C
 #   define BN_MP_ADD_D_C
+#   define BN_S_MP_RAND_CB_C
+#   define BN_S_MP_RAND_SOURCE_C
 #   define BN_MP_PRIME_RANDOM_EX_C
 #endif
 
@@ -761,9 +764,9 @@
 #endif
 
 #if defined(BN_MP_RAND_C)
-#   define BN_S_MP_RAND_SOURCE_PLATFORM_C
 #   define BN_MP_RAND_SOURCE_C
 #   define BN_MP_RAND_DIGIT_C
+#   define BN_S_MP_RAND_SOURCE_C
 #   define BN_MP_ZERO_C
 #   define BN_MP_GROW_C
 #endif
@@ -1168,6 +1171,9 @@
 #if defined(BN_S_MP_MUL_HIGH_DIGS_FAST_C)
 #   define BN_MP_GROW_C
 #   define BN_MP_CLAMP_C
+#endif
+
+#if defined(BN_S_MP_RAND_SOURCE_PLATFORM_C)
 #endif
 
 #if defined(BN_S_MP_REVERSE_C)
